test-pmu: Add initial test
Add the initial version of the test module. It programs every available
general-purpose core hardware performance counter to count instructions
for a tight loop of a million instructions. Typically, three additional
instructions are counted after the loop. The first two are introduced
by the compiler for restoring two registers and the last one is the
MSR write which stops the counter.

The number of available counters are detected based on x86 features
advertised by the kernel.

Signed-off-by: Sandipan Das <sandipan.osd@gmail.com>